requirements.txt
pygame
yt-dlp
websockets
twitchio
gtts
youtubesearchpython
moviepy
python-dotenv